Moved operation to core
