+++ cellst/__init__.py
@@ -4,6 +4,7 @@ __all__ = ['core.orchestrator', 'core.pipeline', 'process', 'segment',
 # Import user-facing classes
 from .core.orchestrator import Orchestrator
 from .core.pipeline import Pipeline
+from .core.arrays import ConditionArray, ExperimentArray
 from .process import Processor
 from .segment import Segmenter
 from .track import Tracker
@@ -11,7 +12,6 @@ from .extract import Extractor
 from .evaluate import Evaluator
 
 # Import important types
-from .core.arrays import ConditionArray, ExperimentArray
 from .utils.utils import ImageHelper
 from .utils._types import Image, Mask, Track, Arr, Same
 from .utils.slurm_utils import SlurmController
